// src/dto/create.ticket.dto.ts
export interface CreateTicketDto {
    temporal: Date;
    count_change: number;
    user_change: string;
    temporal_change: Date;
    ticket: string;
    date_start: Date;
    date_planned: Date;
    attention_type: string;
    idss: string;
    miset: string;
    clase_movimiento: string;
    cod_servicio_descripcion: string;
    cod_servicio_tipo: string;
    tipo_gestion: string;
    nombre_gestion: string;
    area: string;
    service: string;
    service_code: object;
    ddll_reported: string;
    adx_serie_reportada: string;
    description: string;
    company: string;
    division: string;
    name: string;
    code: string;
    email: string;
    segmento_siebel: string;
    clave_sla: string;
    start_attention_am: string;
    end_attention_am: string;
    start_attention_pm: string;
    end_attention_pm: string;
    days_attention: string;
    business_name: string;
    fantasy_name: string;
    rut_commerce: string;
    code_commerce: string;
    region: string;
    city: string;
    locality: string;
    address: string;
    contact_name: string;
    contact_phone: string;
    contact_email: string;
    code_worker: string;
    code_data_commerce: string;
    validar: string;
    temp_valid: Date;
    validador: string;
    asignador: string;
    temp_asig: Date;
    rut_tecnico: string;
    nombre_tecnico: string;
    to: Date;
    ons: Date;
    ce: Date;
    responsable: string;
    rut_responsable: string;
    correo_responsable: string;
    firma_responsable: string;
    comentario_solucion: string;
    status_cambio: boolean;
    status_accesorio: string;
    comentario_accesorio: string;
    cant_ev: number;
    evidencia_1: string;
    evidencia_2: string;
    evidencia_3: string;
    evidencia_4: string;
    evidencia_5: string;
    evidencia_6: string;
    evidencia_7: string;
    evidencia_8: string;
    evidencia_9: string;
    evidencia_10: string;
    id_ticket: string;
    number: string;
    tecnico: string;
    estado_cierre: string;
    adx_categoria_cierre: string;
    adx_codigo_cierre: string;
    adx_equipos_aten_nserie: string;
    adx_equipos_aten_ddll: string;
    adx_equipos_reem_nserie: string;
    adx_equipos_reem_ddll: string;
    adx_fecha_ejecutada: Date;
    cerrar_ticket: boolean;
    solucion: string;
    numero_cue: string;
    fecha_cue: Date;
    contacto: string;
    comentario_contacto: string;
    temporal_contacto: number;
    user_contacto: string;
    status_guia: string;
    fecha_escalamiento: Date;
    comentario_guia: string;
    temporal_status: Date;
    user_status: string;
    solicitud_especial: string;
}